remove image api dependency on docker/distribution/schemaX
